backend/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
Flask==2.3.2
Flask-Cors==4.0.0
flask-expects-json==1.7.0
frozenlist==1.4.0
greenlet==2.0.2
gunicorn==21.2.0
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
multidict==6.0.4
openai==0.27.8
packaging==23.1
poe-api==0.5.0
PySocks==1.7.1
python-socks==2.3.0
quickjs==1.19.2
referencing==0.30.0
requests==2.31.0
rpds-py==0.9.2
tls-client==0.2.1
tqdm==4.65.0
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.1
vercel==0.2.1
websocket-client==1.6.1
Werkzeug==2.3.6
yarl==1.9.2
zope.event==5.0
zope.interface==6.0
